+++ qiskit/quantum_info/states/_measures.py
@@ -16,73 +16,6 @@ A collection of useful quantum information functions for states.
 import numpy as np
 import scipy.linalg as la
 
-from qiskit import QISKitError
-
-
-###############################################################
-# States.
-###############################################################
-
-
-def basis_state(str_state, num):
-    """
-    Return a basis state ndarray.
-
-    Args:
-        str_state (string): a string representing the state.
-        num (int): the number of qubits
-    Returns:
-        ndarray:  state(2**num) a quantum state with basis basis state.
-     Raises:
-        QISKitError: if the dimensions is wrong
-    """
-    n = int(str_state, 2)
-    if num >= len(str_state):
-        state = np.zeros(1 << num, dtype=complex)
-        state[n] = 1
-        return state
-    else:
-        raise QISKitError('size of bitstring is greater than num.')
-
-
-def random_state(num):
-    """
-    Return a random quantum state from the uniform (Haar) measure on
-    state space.
-
-    Args:
-        num (int): the number of qubits
-    Returns:
-        ndarray:  state(2**num) a random quantum state.
-    """
-    x = np.random.random(1 << num)+0.00000001
-    x = -np.log(x)
-    sumx = sum(x)
-    phases = np.random.random(1 << num)*2.0*np.pi
-    return np.sqrt(x/sumx)*np.exp(1j*phases)
-
-
-def projector(state, flatten=False):
-    """
-    maps a pure state to a state matrix
-
-    Args:
-        state (ndarray): the number of qubits
-        flatten (bool): determine if state matrix of column work
-    Returns:
-        ndarray:  state_mat(2**num, 2**num) if flatten is false
-        ndarray:  state_mat(4**num) if flatten is true stacked on by the column
-    """
-    density_matrix = np.outer(state.conjugate(), state)
-    if flatten:
-        return density_matrix.flatten(order='F')
-    return density_matrix
-
-
-###############################################################
-# Measures.
-###############################################################
-
 
 def state_fidelity(state1, state2):
     """Return the state fidelity between two quantum states.
